jobs/nr-garbage-collector/requirements/prod.txt
git+https://github.com/bcgov/namex.git#egg=namex&subdirectory=api
git+https://github.com/bcgov/namex-synonyms-api-py-client.git#egg=swagger_client
git+https://github.com/bcgov/namex-payment-api-py-client.git@dev#egg=openapi_client

cx_Oracle
psycopg2-binary
python-dotenv==0.8.2

Flask==1.1.2
Flask-Migrate
Flask-Script
Flask-Moment
Flask-SQLAlchemy==2.4.1
Flask-RESTplus==0.13.0
Flask-Marshmallow==0.11.0
flask-jwt-oidc>=0.1.5
jsonpickle
pandas
marshmallow==2.19.2
marshmallow-sqlalchemy==0.19.0
pronouncing
requests
toolz
nltk==3.4.5
lxml
inflect
werkzeug==0.16.1
pysolr

